requirements.txt
fastapi
openai
azure-cognitiveservices-speech
python-dotenv
